+++ CHANGES.rst
@@ -10,6 +10,10 @@ Changelog
 2.0a9 (2015-05-04)
 ------------------
 
+- For ``event_listing`` on Collections, ignore the Collection's sorting and use
+  what the event listing's mode defines for sorting.
+  [thet]
+
 - Add support for Collections as data source for calendar and event portlets.
   [thet]